move human agent to pretrained agent folder
